Add / Fix: [WIP] Working docker image with CUDA12.1
Why: cuda was not working on the previous method of config

How:

base image has changed to nvidia's image. no longer using vscode since
that is image bloat

Pytorch is being installed with pip with a cuda requirement. This is
uninstalling the version from conda and reinstalling

Tags: nvidia, cuda, pytorch, pip, conda