feat: Dynamic build and push scripts; conda-lock
version number is picked up from a text file inside of the /docker dir
this is then fed into the build-all.sh and push-all.sh files

we are now using a conda-lock file to ensure proper images. This is
programmatically generated from the environment.yml file inside of
the full container's directory.